backend/requirements.txt
# LiveKit Agents v1.x
livekit-agents>=1.0,<2.0
livekit-plugins-deepgram>=1.0
livekit-plugins-openai>=1.0
livekit-plugins-elevenlabs>=1.0
livekit-plugins-silero>=1.0
livekit-plugins-turn-detector>=1.0

# RAG stack
langchain>=0.3,<0.4
langchain-openai>=0.2,<0.4
langchain-community>=0.3,<0.4
langchain-chroma>=0.1
chromadb>=0.5
pypdf>=5.0

# Misc
python-dotenv>=1.0
tiktoken>=0.7
numpy>=1.26
